+++ okdata/cli/__main__.py
@@ -11,6 +11,7 @@ from okdata.cli.commands.permissions import PermissionsCommand
 from okdata.cli.commands.pipelines import Pipelines
 from okdata.cli.commands.pubreg import PubregCommand
 from okdata.cli.commands.status import StatusCommand
+from okdata.cli.commands.teams import TeamsCommand
 
 
 def main():
@@ -71,6 +72,7 @@ def get_command_class(argv):
         "pipelines": Pipelines,
         "pubreg": PubregCommand,
         "status": StatusCommand,
+        "teams": TeamsCommand,
     }
     return commands.get(argv[1], False)